re-arrange components
