+++ baby-gru/src/components/card/MoorhenMoleculeCard.tsx
@@ -1,35 +1,19 @@
 import { useEffect, useState, useRef, useReducer, useCallback, useImperativeHandle, forwardRef } from 'react';
-import { Card, Row, Col, Accordion, Stack, Button, FormSelect, Form } from "react-bootstrap";
-import { doDownload, rgbToHex, sequenceIsValid } from '../../utils/MoorhenUtils';
+import { Card, Row, Col, Accordion, Stack, Button } from "react-bootstrap";
+import { doDownload, sequenceIsValid, representationLabelMapping } from '../../utils/MoorhenUtils';
 import { isDarkBackground } from '../../WebGLgComponents/mgWebGL'
 import { MoorhenSequenceViewer } from "../sequence-viewer/MoorhenSequenceViewer";
 import { MoorhenMoleculeCardButtonBar } from "../button-bar/MoorhenMoleculeCardButtonBar"
 import { MoorhenLigandList } from "../list/MoorhenLigandList"
-import { Chip, FormGroup, Popover, hexToRgb } from "@mui/material";
+import { Chip, FormGroup, hexToRgb } from "@mui/material";
 import { getNameLabel } from "./cardUtils"
 import { moorhen } from "../../types/moorhen";
 import { webGL } from "../../types/mgWebGL";
 import { AddOutlined, DeleteOutlined, FormatColorFillOutlined } from '@mui/icons-material';
-import { SliderPicker } from "react-color";
-import { MoorhenColourRules } from '../modal/MoorhenColourRules';
-
-const labelMapping = {
-    rama: "Rama.",
-    rotamer: "Rota.",
-    CBs: "Bonds",
-    CAs: "C-As",
-    CRs: "Ribbons",
-    CDs: "Cont. dots",
-    MolecularSurface: "Surf.",
-    gaussian: "Gauss.",
-    ligands: "Ligands",
-    DishyBases: "Bases",
-    VdwSpheres: "Spheres",
-    allHBonds: "H-Bonds"
-}
+import { MoorhenAddCustomRepresentationCard } from "./MoorhenAddCustomRepresentationCard"
+import { MoorhenModifyColourRulesCard } from './MoorhenModifyColourRulesCard';
 
 const allRepresentations = [ 'CBs', 'CAs', 'CRs', 'ligands', 'gaussian', 'MolecularSurface', 'DishyBases', 'VdwSpheres', 'rama', 'rotamer', 'CDs', 'allHBonds' ]
-const customRepresentations = [ 'CBs', 'CAs', 'CRs', 'ligands', 'gaussian', 'MolecularSurface', 'DishyBases', 'VdwSpheres' ]
 
 interface MoorhenMoleculeCardPropsInterface extends moorhen.Controls {
     dropdownId: number;
@@ -469,7 +453,7 @@ export const MoorhenMoleculeCard = forwardRef<any, MoorhenMoleculeCardPropsInter
                             <FormatColorFillOutlined/>
                         </Button>
                     </Col>
-                    <MoorhenColourRules molecules={props.molecules} anchorEl={addColourRulesAnchorDivRef} isDark={props.isDark} urlPrefix={props.urlPrefix} glRef={props.glRef} commandCentre={props.commandCentre} molecule={props.molecule} showColourRulesToast={showColourRulesModal} setShowColourRulesToast={setShowColourRulesModal} windowHeight={props.windowHeight} windowWidth={props.sideBarWidth}/>
+                    <MoorhenModifyColourRulesCard molecules={props.molecules} anchorEl={addColourRulesAnchorDivRef} isDark={props.isDark} urlPrefix={props.urlPrefix} glRef={props.glRef} commandCentre={props.commandCentre} molecule={props.molecule} showColourRulesToast={showColourRulesModal} setShowColourRulesToast={setShowColourRulesModal} windowHeight={props.windowHeight} windowWidth={props.sideBarWidth}/>
                 </Row>
                 <Row style={{display: 'flex'}}>
                     <Col  style={{ width:'100%', height: '100%' }}>
@@ -489,7 +473,7 @@ export const MoorhenMoleculeCard = forwardRef<any, MoorhenMoleculeCardPropsInter
                         <Button style={{ paddingTop: '0.25rem', paddingBottom: '0.25rem' }} variant='light' onClick={() => setShowCreateCustomRepresentation((prev) => {return !prev})}>
                             <AddOutlined/>
                         </Button>
-                        <CreateCustomRepresentationMenu molecule={props.molecule} anchorEl={addCustomRepresentationAnchorDivRef} show={showCreateCustomRepresentation} setShow={setShowCreateCustomRepresentation}/>
+                        <MoorhenAddCustomRepresentationCard molecule={props.molecule} anchorEl={addCustomRepresentationAnchorDivRef} show={showCreateCustomRepresentation} setShow={setShowCreateCustomRepresentation}/>
                     </Col>
                 </Row>                
             <Accordion alwaysOpen={true} defaultActiveKey={['sequences']}>
@@ -598,49 +582,9 @@ const RepresentationCheckbox = (props: {
     return <Chip
                 style={chipStyle}
                 variant={"outlined"}
-                label={`${labelMapping[props.repKey]}`}
+                label={`${representationLabelMapping[props.repKey]}`}
                 onClick={handleClick}
             />
-}
-
-const CreateCustomRepresentationMenu = (props: {
-    setShow: React.Dispatch<React.SetStateAction<boolean>>;
-    show: boolean; anchorEl: React.RefObject<HTMLDivElement>;
-    molecule: moorhen.Molecule
-}) => {
-
-    const cidFormRef = useRef<HTMLInputElement | null>(null)
-    const styleSelectRef = useRef<HTMLSelectElement | null>(null)
-    const [colour, setColour] = useState<{ r: string; g: string; b: string; }>({r: '150', g:'100', b:'50'})
-
-    const handleCreateRepresentation = useCallback(() => {
-        props.molecule.addRepresentation(styleSelectRef.current.value, cidFormRef.current.value, true, rgbToHex(parseInt(colour.r), parseInt(colour.g), parseInt(colour.b)))
-        props.setShow(false)
-    }, [colour])
-
-    return <Popover
-                onClose={() => props.setShow(false)}
-                open={props.show}
-                anchorEl={props.anchorEl.current}
-                anchorOrigin={{ vertical: 'center', horizontal: 'center' }}
-                transformOrigin={{ vertical: 'center', horizontal: 'center', }}
-                sx={{'& .MuiPaper-root': {backgroundColor: '#e8e8e8', marginTop: '0.1rem'}}}
-            >
-            <Stack gap={2} direction='horizontal' style={{width: '25rem', margin: '0.5rem'}}>
-                <FormSelect ref={styleSelectRef} size="sm" defaultValue={'Bonds'}>
-                    {customRepresentations.map(key => {
-                        return <option value={key} key={key}>{labelMapping[key]}</option>
-                    })}
-                </FormSelect>
-                <Form.Control ref={cidFormRef} size="sm" type='text' placeholder={'CID selection'} style={{width: "100%"}}/>
-                <div style={{width: '100%', textAlign: 'center'}}>
-                <SliderPicker color={colour} onChange={(color) => {setColour({r: color.rgb.r, g: color.rgb.g, b: color.rgb.b})}}/>
-                </div>
-                <Button onClick={handleCreateRepresentation}>
-                    Create
-                </Button>
-            </Stack>
-        </Popover>
 }
 
 const CustomRepresentationChip = (props: {
@@ -669,7 +613,7 @@ const CustomRepresentationChip = (props: {
     return <Chip
         style={{marginLeft: '0.1rem', marginBottom: '0.1rem', borderColor: `rgb(${r}, ${g}, ${b})`, backgroundColor: `rgba(${r}, ${g}, ${b}, ${representationIsVisible ? 0.5 : 0.1})`}}
         variant={"outlined"}
-        label={`${labelMapping[representation.style]} ${representation.cid}`}
+        label={`${representationLabelMapping[representation.style]} ${representation.cid}`}
         deleteIcon={<DeleteOutlined/>}
         onDelete={() => {
             molecule.removeRepresentation(representation.uniqueId)
